update product migration

+++ product-api/app/Database/Migrations/2025-07-28-101238_CreateProductTable.php
@@ -26,7 +26,7 @@ class CreateProductTable extends Migration
                 'null' => false,
             ],
             'price' => [
-                'type' => 'DECIMAL',
+                'type' => 'INT',
                 'constraint' => 11,
                 'null' => false
             ],
